+++ src/hooks/useDragBlock.ts
@@ -1,39 +1,35 @@
-import { useState } from 'react';
+import React, { Dispatch, SetStateAction } from 'react';
 
-function useDragBlock() {
-  //서버처리도 해주고
-  //클라 색칠도 해주고
-  const [draged, setDraged] = useState('');
+function useDragBlock(
+  isDragging: boolean,
+  handleDragState: (isDragging: boolean) => void,
+  setStart: Dispatch<SetStateAction<string>>,
+  setEnd: Dispatch<SetStateAction<string>>,
+) {
+  const onMouseDown = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (e.target instanceof HTMLDivElement) {
+      setStart(e.target.id);
+    }
+    handleDragState(true);
+  };
 
-  const onDragStart = (e: React.DragEvent<HTMLDivElement>) => {
-    console.log('drag');
-    const img = new Image();
-    e.dataTransfer.setDragImage(img, 0, 0);
-    if (e.target instanceof HTMLElement) {
-      e.target.style.cursor = 'pointer';
+  const onMouseEnter = (e: React.MouseEvent<HTMLDivElement>) => {
+    if (isDragging) {
+      if (e.target instanceof HTMLElement) {
+        setEnd(e.target.id);
+      }
     }
   };
 
-  const onDragOver = (e: React.DragEvent<HTMLDivElement>) => {
-    console.log('dragIng');
-    if (e.target instanceof HTMLElement) {
-      e.target.style.cursor = 'pointer';
-      setDraged('plan');
-    }
-  };
-
-  const onDragEnd = (e: React.DragEvent<HTMLDivElement>) => {
-    if (e.target instanceof HTMLElement) {
-      console.log('dragEnd');
-      e.target.style.cursor = 'pointer';
-    }
+  const onMouseUp = () => {
+    handleDragState(false);
+    //서버처리
   };
 
   return {
-    onDragStart,
-    onDragOver,
-    onDragEnd,
-    draged,
+    onMouseDown,
+    onMouseEnter,
+    onMouseUp,
   };
 }
